requirements.txt
requests
pyperclip
rich>=13.3.1
prompt_toolkit>=3.0
sseclient-py>=1.7.2
tiktoken
packaging
poe-api
python-i18n